pyproject.toml
[tool.poetry]
name = "bingoset"
version = "0.1.0"
description = ""
authors = ["Akshay Ashok <aks28id@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
bingoset = "bingoset.main:app"

[tool.poetry.dependencies]
python = "^3.7"
typer = {extras = ["all"], version = "^0.1.1"}
Pillow = "^7.1.2"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
